Much smoother scrolling
The drawRect method now does a synchronous RPC to the core, which is
really the way Cocoa wants to do it. Scrolling no longer causes
needsDisplay, which also results in a lot less redrawing.

Experimentation shows that we can probably get rid of proactively
sending the lines on update, as it's very hard to predict the area
redrawn (due to scrolling).

Since there's synchronous RPC, this commit also contains some logic to
do those.