Refactor front-end/back-end protocol to allow tabs
A significant restructuring, motivated to allow editing of more than
one tab at a time. With this commit, the front-end/back-end protocol
supports multiple tabs, but that functionality is not implemented in
the front-end yet.

The RPC mechanism has been reworked to be much more like JSON-RPC 2.
Thus, this patch closes #29.

This also closes #44, which describes the motivation for these
changes.