Fix Autoresize/Constraint conflict
Fix a conflict with constraints generated from the EditView’s
auto-resize mask. These extraneous constraints cause Xi edit windows
not to be scrollable or resizable on some versions of Xcode. The effect
of this conflict depends on which of the conflicting constraints are
discarded, which seems to somewhat unpredictable. This change marks the
EditView as translatesAutoresizingMaskIntoConstraints=NO.