Incremental update protocol
Starting to implement the new protocol described in #152.

As of this commit, the protocol should be in place, and functionality
should be basically unchanged, but the implementation is not very
optimized. In particular, updates just send the scroll area (with
slop), invalidating everything else, rather than actually sending a
small delta.

In addition, there are a whole bunch of optimizations not done yet,
for example style spans are stored as (old-style) view::Style objects,
and looked up in the style map one-at-a-time while rendering. A far
more efficient approach would be to use a similar style id system in
the plug-in protocol, and then basically plumb style id's through.

That said, all these optimizations should be possible without change
to the update protocol, and I'm reasonably satisfied that will be
stable now.