Reuse empty views when opening a new document
This is a naive implementation of this functionality. In the view
controller, we intercept all calls to `openDocument:`, and check to see
if our current document is empty. If it is, we save a reference to it in the app
delegate. We check this reference when opening a document, and reuse
existing windows if appropriate.

This does not currently cover the 'recent documents' menu, which has a
different implementation; it calls the application delegate's
application:openFile: method.