Stop treating document as an optional
It should be expected that a given view / viewcontroller always has a
corresponding backing document. If it doesn't, we currently fail
silently, which makes debugging more difficult.

The EditView also no longer needs to have its own reference to the
document. This can instead be exposed via the datasource protocol (which
may be poorly named).