Set "opaque" on EditView
Setting EditView to be opaque should improve performance a bit because
the region behind the view doesn't need to be cleared. This setting
also causes text to be rendered with subpixel aliasing when
experimenting with changing EditView to be layer-backed.

Also contains a couple of minor cleanups, including updating the
commented-out dirty rect visualization, and getting rid of a Swift
warning on String.characters.count.