Add font style spans
Also correctly handles flushing of font cache. This is tuned ok for
code editing but is small for rich text / CJK / etc., not to mention
the extra cache pressure from subpixel positioning.

Adds fake italics, which is a reasonably deep change because it needs
to be part of the key for glyph cache lookup.