Cleanups
Removed a bunch of code that was no longer needed, especially
GutterView and the draw() method of EditView (all rendering is now
done by TextPlane). Also got rid of XiScrollView as we no longer need
to disable responsive scrolling.

Also handled the case where the CTLine getters returned nil, which
should help with the observed crashes.