Change locking strategy for LineCache
We want to allow the LineCache to be locked mutably for an extended
period so we can store associated data (TextLine entries and probably
caret positions). This patch introduces a MutexGuard very similar to
Rust's, and also a LineCacheLocked that is an instance, so mutations
to the state are safely guarded by the mutex.