Remove gutter from line cache
The gutter invalidation logic is different from regular lines.  It
should only be invalidated when a newline is inserted & that newline
insertion invalidates all following lines.  This is in contrast to other
things using the line cache which only invalidate on line changes &
never newlines.  In fact, the gutter is invalidated whenever the cursor
changes lines too.

This fixes the bug where adding a newline in the middle of a document
has an off-by-1 error in the gutter line count.