More rational view reuse, fewer redundant redraws
There was a small bug where we were incorrectly comparing new font names
to old ones, which caused us to unnecessarily redraw all views whenever
a new view was opened or a config changed.

We were also apparently relying on this behaviour to redraw the content
of reused views, which we now do explicitly.

- fixes #130