Scroll to the the old visible area after find bar is shown (#225)
* Scroll to the current origin after find bar is shown

* Rename the var

Because the old origin of the reat has no different to the new origin. It would be better to rename the var to `origin`, to avoid any misunderstanding

* Change to adjust scrollView instead of the frame of it

* Fix: the scroll is down if open view while view is opened

We should onlly adjust the scroll view if the find bar is not visible when the open view is triggered.

* Remove extra empty line

* Fixed: scroll is not reset when te find bar is closed

* Fix layout error when toggling replace bar

This commit fix two issus with replace penal.

1. Can't scroll to first line when replace panel is shown.
2. The view point changed after toggling replace bar

The first issue is caused by the height of the repalce panel is ignored when calcualte offset. I fixed it by calculate offset after showing replace panel.

Issue two is causeb by the offset set in previous `openFind()` are accumulated. I fixed it by removing the offset in view point set in previous `openFind()` before setting the new view point.

* Fix view point is not reset to original value

replacePanel.isHidden should not be set before view point is reset, otherwise, the reset value will be incorrect

* Extract declarations to the top of function