Unify & simplify some view geometry calculations
This unifies the 'offset to line' calculations, and removes
the separate storage of a scroll offset in the EditView.

This tries to have  the y-offset to line number math in one place.
Additionally we don't track the scroll offset independently, and
assume that the view's bounds are up to date when we need them.