Add context menu on EditContainerView
Implement a default context menu (for any event) on EditContainerView. The menu itself is generated in EditViewController so that it shares the same actions as other menu items and can be easily re-configured.

In addition, overriding menu(for:) in EditContainerView allows macOS to determine when/where to present the menu regardless of the user’s right click preferences.

Finally, having the menu on EditContaierView ensure it captures all possible events that could trigger a context menu because it is the top most view of the editor’s hierarchy.

Resolves: #306